Sol 6-4: Use max count instead of static number.
Print the sorted list of words by counts on input using a
dynamically tracked maximum count instead of the method in the previous
commit which just uses a number related to the test input.
Now this method should work for arbitrary input instead of just input
similar to the test input.