dcl: rename complete dcl_main.c to dcl_full.c
Reorganizing a few different things. The name of the dcl implementation
that included all functions in one file was "dcl_main.c" but is now
"dcl_full.c". This is so dcl_main.c can be used as the name of the split
dcl implementation file that contains the main() function. dcl() and
dirdcl() will stay in dcl.c. undcl will have its own main in
undcl_main.c and link to getch(), gettoken(), dcl(), etc.