Exercise 5-18: dcl: use int for missing type.
A naive addition to the dcl application where a default type of 'int' is
used instead of throwing an error for missing type. The text says the
dcl example is for explanatory purposes and so are these solutions;
they're not robust against deeper errors.
Start moving test code to a test sub-directory.