Solution 5-19: undcl: don't write extra ()'s.
This is an initial attempt at the solution for Exercise 5-19. The
exercise asks the reader to prevent undcl from writing out unnecessary
parentheses (()'s).
The undcl application will convert a description of a declaration
(written in a restricted grammar) to a C-style declaration.
By default, the undcl example from the text will write ()'s
around the name or output thus far for every '*' it finds.
This solution
considers this a slightly modified approach where is only necessary
if the order of operations requires the ()'s to bind the '*'
to the name instead of another construct based ()'s and []'s seen on the
input.
This initial solution counts the number of ()'s and []'s (parentheses
and brackets) seen so far and only writes ()'s for any *'s were the ()'s
and []'s count is greater than 0.
The idea is for relatively simple input this modification will prevent
unnecessary ()'s because only binding to ()'s and []'s is considered
necessary to add ()'s to the output.