Solution 5-20: Find function argument in C declaration.
This is another intermediate step towards a more complete exercise 5-20
solution. If the C declaration input contains a function argument this
version will find it and print it with the rest of the English text
explanation of the C declaration. This is more of a idea demonstration
that a complete function argument solution: only one argument is
supported, the argument type is not checked and no qualifiers are
supported. Eventually there will need to be a breakdown of how this
change affects the recursive decent parser used by the DCL example from
the text.
This solution was worked out in the _arg.c/.h source files and moved to
the _ex5-20.c/.h source files.