Exercise 5-20: Better function argument support.
There are a few changes in this single commit. The complete string
representing the function argument is now passed as-is to the output,
instead of stopping at just the first word. Even though the complete
function argument string is passed now, the English description of the C
declaration is now compromised and harder to read. Maybe this can be
fixed later.
Also includes minor changes for head file names so that the name matches
the related exercise number.