Solution 6-1: Add underscore handling to getword().
This commit is a cleaner version of the solution to the first part of
exercise 6-1, add underscore support to getword(). This is a very short
change for bare minimum functionality. This solution will also include
test inputs and test outputs in follow-on commits.
This commit also has a much more expanded description of the original
getword() function in the commend header for ex6_1_getword().