Solution 6-3: Dynamic array for line numbers.
Not sure why this works yet, if it works. The idea was replace the
static array in the tnode definition with a pointer to a place in the
heap where line numbers for words would be stored. There is a problem
with this: getting functions, struct pointer passing and realloc() to all
play nice. This version seems to run without crashing so I'm stopping
here for now. But left a TODO to make this code better. The allocation
should be tighter for example, and the bugs with realloc() should be
investigated.