Solution 7-4: Start of solution using local scanf().
The local scanf() the text is looking for is called minscanf(), similar
to the printf() example from Section 7.3. This commit has a lot of
unnecessary code and debug code including a practice function
protoscanf(). The problem was with the call to va_arg(). The 2nd
parameter needs to be 'int *' for an integer argument, which this
comment has. Major cleanup and more parts of the exercise still needed.