fflush: Fix bug with wrong buffer size flushed.
The problem was BUFSIZ was being defined by something included into the
build, causing fflush to write out more than the size of the file buffer
specified in the FILE * argument. This commit fixes that. This commit
also disables flushing default file buffers in the NULL call case
because the implementation was incomplete and likely incorrect.